// src/app/shared/models/call-contact.ts
export class CallContact{
    id: string;
    call_id: string;
    contact_id: string;
    required: string;
    accept_status: string;
    date_modified: Date;
    deleted: boolean;
}

export class CallContactJSon{
    id: string;
    callId: string;
    contactId: string;
    required: string;
    acceptStatus: string;
    dateModified: Date;
    deleted: boolean;
}